requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
openai==0.28.0
nltk==3.8.1
spacy==3.6.1
PyPDF2==3.0.1
pdfplumber==0.10.2
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1
asyncio==3.4.3
aiofiles==23.2.1
pymongo==4.5.0
